posix thread implementation

+++ yu/src/core/crc32.cpp
@@ -1,3 +1,4 @@
+#include "platform.h"
 #include "crc32.h"
 #include "type.h"
 
@@ -14,7 +15,7 @@ namespace yu
 
 #ifdef CREATE_CRC_TABLE
 
-static unsigned long crctable[256];
+static unsigned int crctable[256];
 
 /*
  Generate a table for a byte-wise 32-bit CRC calculation on the polynomial:
@@ -43,7 +44,7 @@ static unsigned long crctable[256];
 
 void make_crc_table( void ) {
 	int i, j;
-	unsigned long c, poly;
+	unsigned int c, poly;
 	/* terms of polynomial defining this crc (except x^32): */
 	static const byte p[] = {0,1,2,4,5,7,8,10,11,12,16,22,23,26};
 	
@@ -54,7 +55,7 @@ void make_crc_table( void ) {
 	}
 	
 	for ( i = 0; i < 256; i++ ) {
-		c = (unsigned long)i;
+		c = (unsigned int)i;
 		for ( j = 0; j < 8; j++ ) {
 			c = ( c & 1 ) ? poly ^ ( c >> 1 ) : ( c >> 1 );
 		}
@@ -67,7 +68,7 @@ void make_crc_table( void ) {
 /*
  Table of CRC-32's of all single-byte values (made by make_crc_table)
  */
-static unsigned long crctable[256] = {
+static unsigned int crctable[256] = {
 	0x00000000L, 0x77073096L, 0xee0e612cL, 0x990951baL,
 	0x076dc419L, 0x706af48fL, 0xe963a535L, 0x9e6495a3L,
 	0x0edb8832L, 0x79dcb8a4L, 0xe0d5e91eL, 0x97d2d988L,
@@ -136,18 +137,18 @@ static unsigned long crctable[256] = {
 
 #endif
 
-void CRC32_InitChecksum( unsigned long &crcvalue ) 
+void CRC32_InitChecksum( unsigned int &crcvalue )
 {
 	crcvalue = CRC32_INIT_VALUE;
 }
 
-void CRC32_Update( unsigned long &crcvalue, const u8 data ) 
+void CRC32_Update( unsigned int &crcvalue, const u8 data )
 {
 	crcvalue = crctable[ ( crcvalue ^ data ) & 0xff ] ^ ( crcvalue >> 8 );
 }
 
-void CRC32_UpdateChecksum(unsigned long &crcvalue, const void *data, size_t length) {
-	unsigned long crc;
+void CRC32_UpdateChecksum(unsigned int &crcvalue, const void *data, size_t length) {
+	unsigned int crc;
 	const unsigned char *buf = (const unsigned char *) data;
 
 	crc = crcvalue;
@@ -157,12 +158,12 @@ void CRC32_UpdateChecksum(unsigned long &crcvalue, const void *data, size_t leng
 	crcvalue = crc;
 }
 
-void CRC32_FinishChecksum( unsigned long &crcvalue ) {
+void CRC32_FinishChecksum( unsigned int &crcvalue ) {
 	crcvalue ^= CRC32_XOR_VALUE;
 }
 
-unsigned long CRC32_BlockChecksum(const void *data, size_t length) {
-	unsigned long crc;
+unsigned int CRC32_BlockChecksum(const void *data, size_t length) {
+	unsigned int crc;
 
 	CRC32_InitChecksum( crc );
 	CRC32_UpdateChecksum( crc, data, length );
@@ -170,7 +171,7 @@ unsigned long CRC32_BlockChecksum(const void *data, size_t length) {
 	return crc;
 }
 
-unsigned long CRC32(const void *data, size_t length)
+unsigned int CRC32(const void *data, size_t length)
 {
 	return CRC32_BlockChecksum(data, length);
 }
